deepspeed/requirements.txt
nbox[serving]
transformers
datasets
pillow==8.2
evaluate
deepspeed
accelerate
scikit-learn